% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_heatmap.R
\name{rule_heatmap}
\alias{rule_heatmap}
\title{Create a Heatmap for Association Rules}
\usage{
rule_heatmap(
  rules,
  metric = "confidence",
  graph_title = "",
  low_color = "lightblue",
  high_color = "navy",
  include_zero = FALSE
)
}
\arguments{
\item{rules}{An object of class \code{rules}, typically generated by the \code{apriori} function from the \code{arules} package.}

\item{metric}{A character string specifying the metric to use for coloring the heatmap.
Must be one of \code{"confidence"}, \code{"support"}, or \code{"lift"}. Defaults to \code{"confidence"}.}

\item{graph_title}{A character string specifying the title of the graph.
Defaults to an empty string (\code{""}).}

\item{low_color}{A valid R color or hex color code for the lower bound of the gradient.
Defaults to \code{"lightblue"}.}

\item{high_color}{A valid R color or hex color code for the upper bound of the gradient.
Defaults to \code{"navy"}.}

\item{include_zero}{A logical value indicating whether to include zero values for missing antecedent-consequent combinations.
Defaults to \code{FALSE}.}
}
\value{
A \code{ggplot} object representing the heatmap visualization of the association rules.
}
\description{
The \code{rule_heatmap} function generates a heatmap visualization of association rules,
showing relationships between antecedents and consequents based on a specified metric.
}
\examples{
library(arules)
data(Groceries)

# Generate rules
rules <- apriori(Groceries, parameter = list(supp = 0.01, conf = 0.5, target = "rules"))

# Create a heatmap of the rules using confidence as the metric
rule_heatmap(rules, metric = "confidence", graph_title = "Confidence Heatmap")

# Create a heatmap of the rules using lift as the metric
rule_heatmap(
  rules, 
  metric = "lift", 
  graph_title = "Lift Heatmap", 
  low_color = "white", 
  high_color = "red"
)

}
