% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_heatmap.R
\name{rule_heatmap}
\alias{rule_heatmap}
\title{Create a Heatmap for Association Rules}
\usage{
rule_heatmap(
  rules,
  metric = "confidence",
  graph_title = "",
  low_color = "lightblue",
  high_color = "navy",
  include_zero = FALSE
)
}
\arguments{
\item{rules}{An object of class \code{rules} from the \code{arules} package.}

\item{metric}{A character string specifying the metric to use for coloring the heatmap.
Must be one of \code{"confidence"}, \code{"support"}, or \code{"lift"}. Defaults to \code{"confidence"}.}

\item{graph_title}{A character string specifying the title of the graph.
Defaults to an empty string (\code{""}).}

\item{low_color}{A valid R color or hex color code for the lower bound of the gradient.
Defaults to \code{"lightblue"}.}

\item{high_color}{A valid R color or hex color code for the upper bound of the gradient.
Defaults to \code{"navy"}.}

\item{include_zero}{A logical value indicating whether to include zero values for missing antecedent-consequent combinations.
Defaults to \code{FALSE}.}
}
\value{
A \code{ggplot} object representing the heatmap visualization of the association rules.
}
\description{
Generates a heatmap visualization of association rules,
showing relationships between antecedents and consequents based on a specified metric.
}
\examples{
library(arules)
library(tidyr)
data(BrookTrout)

# Discretise data
discrete_bt <- dtize_df(BrookTrout, cutoff="median")

# Generate rules
rules <- apriori(
  discrete_bt, 
  parameter = list(supp = 0.01, conf = 0.5, target = "rules"), 
  appearance = list(rhs="eDNAConc=high")
)

# Subset ruleset (too many rules won't fit on the heatmap)
rules <- rules \%>\%
  subset(!is.redundant(., measure = "confidence")) \%>\%
  subset(is.significant(., alpha = 0.05)) \%>\%
  sort(by = c("confidence", "lift", "support"))

# Create a heatmap of the rules using confidence as the metric
rule_heatmap(
  rules, 
  metric = "confidence", 
  graph_title = "Confidence Heatmap"
)

# Create a heatmap of the rules using lift as the metric
rule_heatmap(
  rules, 
  metric = "lift", 
  graph_title = "Lift Heatmap", 
  low_color = "#D4A221", 
  high_color = "darkgreen"
)

}
